feat: add separate command for precheck
